backend/pyproject.toml
[project]
name = "researchinstituteofsquids"
version = "0.1.0"
description = "An agentic research institute where AI agents read papers, form hypotheses, debate, run experiments, and build shared knowledge graphs."
requires-python = ">=3.12"
dependencies = [
    # API
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",

    # LLM & Orchestration
    "langgraph>=0.2",
    "langchain-openai>=0.2",
    "openai>=1.0",

    # RAG & Indexing
    "llama-index-core>=0.11",
    "llama-index-vector-stores-postgres>=0.2",
    "llama-index-readers-file>=0.2",
    "llama-index-embeddings-openai>=0.2",

    # Knowledge Graph
    "neo4j>=5.0",

    # Database & Memory
    "asyncpg>=0.29",
    "pgvector>=0.3",
    "sqlalchemy[asyncio]>=2.0",
    "hindsight-all>=0.1",

    # Data Models & Config
    "pydantic>=2.0",
    "pydantic-settings>=2.0",

    # CLI
    "typer>=0.12",
    "rich>=13.0",

    # Sandbox
    "docker>=7.0",

    # Ingestion
    "httpx>=0.27",
    "pymupdf>=1.24",
    "beautifulsoup4>=4.12",

    # Search
    "tavily-python>=0.4",
    "arxiv>=2.0",

    # Utilities
    "python-dotenv>=1.0",
]

[project.scripts]
squid = "src.cli.main:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]
